refactor: changed_work_image method - ui로 이동
